Use Py_ssize_t instead of long to handle 64bit indices

+++ scipy/sparse/csgraph/_traversal.pyx
@@ -400,8 +400,8 @@ cdef unsigned int _breadth_first_undirected(
     return i_nl
 
 cdef struct DfsStackEntry:
-    long node
-    long index
+    Py_ssize_t node
+    Py_ssize_t index
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
